pyproject.toml
[tool.poetry]
name = "fastapi-application"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
bcrypt = "^4.0.1"
python-jose = "^3.3.0"
fastapi = "^0.95.2"
uvicorn= {extras = ["standard"], version = "^0.20.0"}
python-multipart = "^0.0.6"
pymongo = {extras = ["srv"], version = "^4.3.3"}
orjson = "^3.9.1"
python-dotenv = "^1.0.0"
boto3 = "^1.26.153"
pydantic = "^1.10.9"
requests = "^2.31.0"
langchain = "^0.0.221"
pinecone-client = "^2.2.2"
openai = "^0.27.8"
bs4 = "^0.0.1"
google-search-results = "^2.4.2"
asyncio = "^3.4.3"
sse-starlette = "^1.6.1"
cohere = "^4.17.0"



[tool.poetry.group.dev.dependencies]
types-python-jose = "^3.3.4.7"
ruff = "^0.0.271"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 
